calling server functions, mocks now works
Mocking:
- isJest.js > This will become useful if/when we want to run tests.
- getMocks.js > This is where we can provide mock data/behavior when developing locally
- polyfillScriptRun.js > Necessary for runGas.js to work locally and in GAS env
- runGas.js > Wrapper function for calling server functions. Works locally, too.

Examples:
The UserPreferences page has a save button that calls for getUserPreferences
and logs the response. Thanks to the above, this works locally, returning a
fake user name 'John Mock', while in GAS it returns 'John Server'.

Next:
We'll try to integrate the PropertiesService on the server instead of just
returning an object for John Server