Fix for broken view API response
After some hair pulling it turns out that returning a date from the GAS
server to the client may be a no-no.
See this [stackoverflow](https://stackoverflow.com/questions/56895811/google-script-run-executes-but-it-returns-null-while-code-gs-function-does-hav) link.
My issue seems to have been caused by including a `producedAt` Date
property in the `QueryResult` object that I create when querying the datasource.