Script Props for User Preferences
If deploying the app to `"executeAs": "USER_DEPLOYING"` in the
manifest, that means the User Properties are basically useless. Or
at the very least, they do not satisfy the assumption that the
accessing user's User Properties can be used to store things that
only pertain to them.

Therefore we've switched the storage/persistence mechanism to use
the Script Properties instead. We'll see if this stands the test
of time, but it is necessary for the moment